LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/document-write/module-tla-import-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected keyword 'import'

Harness Error (TIMEOUT), message = null

TIMEOUT document.write in an imported module Test timed out
